Write a program that divides two real numbers (FLOAT type) entered by the user.  Take care of the appropriate formatting of the result and messages guiding the user  through the data entry, and also protect the program against division by the number zero, i.e. if the second number is zero, the program should not calculate the result,  but print a warning message and stop.
 Write a program that divides two real numbers (FLOAT type) entered by the user. 
Take care of the appropriate formatting of the result and messages guiding the user 
through the data entry, and also protect the program against division by the number zero,
i.e. if the second number is zero, the program should not calculate the result, 
but print a warning message and stop.