write a program that calculates the surface area and volume of a cube for a user-specified length of the cube's edge. The program should be protected against entering incorrect data, i.e. the length of the cube's edge should be positive. Otherwise, the program should write a warning message and stop. Use the IF-ELSE statement
write a program that calculates the surface area and volume of a cube for
a user-specified length of the cube's edge. The program should be protected against
entering incorrect data, i.e. the length of the cube's edge should be positive.
Otherwise, the program should write a warning message and stop. Use the IF-ELSE statement