Write a program that calculates the area of a trapezoid for positive numbers a, b,h (lengths of the bases and height) given by the user, displaying an appropriate message on the screen. The program should be protected against entering incorrect data, i.e. the lengths a, b, h should be positive.  Otherwise, the program should write a warning message and stop. Use the IF-ELSE instruction
Write a program that calculates the area of a trapezoid for positive
numbers a, b,h (lengths of the bases and height) given by the user, displaying an
appropriate message on the screen. The program should be protected against entering
incorrect data, i.e. the lengths a, b, h should be positive. 
Otherwise, the program should write a warning message and stop. Use the IF-ELSE instruction